+++ psycop/common/sequence_models/tasks/__init__.py
@@ -1,9 +1,9 @@
-from .task_registry import create_behrt, create_encoder_for_clf
+from .task_registry import create_behrt, clf_encoder
 from .tasks import BEHRTForMaskedLM, EncoderForClassification
 
 __all__ = [
     "BEHRTForMaskedLM",
     "EncoderForClassification",
     "create_behrt",
-    "create_encoder_for_clf",
+    "clf_encoder",
 ]